rename `markers.py` module
