+++ tests/fixtures/core/__init__.py
@@ -5,9 +5,9 @@ from tests.fixtures.core.api import as_user
 from tests.fixtures.core.current_user import _set_current_user
 from tests.fixtures.core.factory import factory
 from tests.fixtures.core.factory import mixer
-from tests.fixtures.core.markers import _pause_auditlog_or_set_current_user
-from tests.fixtures.core.markers import _pause_dashamail
-from tests.fixtures.core.markers import _pause_tags
+from tests.fixtures.core.pause_slow_background_jobs import _pause_auditlog_or_set_current_user
+from tests.fixtures.core.pause_slow_background_jobs import _pause_dashamail
+from tests.fixtures.core.pause_slow_background_jobs import _pause_tags
 from tests.fixtures.core.send_mail import send_mail
 from tests.fixtures.core.timezone import kamchatka_timezone
