+++ tests/fixtures/core/pause_slow_background_jobs.py
